requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
requests==2.32.4
pydantic==2.11.7
pydantic-ai==0.0.14
python-dotenv==1.0.0
setuptools>=65.0.0 